requirements.txt
watchdog==4.0.0
Pillow==10.4.0
numpy<2
onnxruntime==1.17.1
rembg==2.0.57
opencv-python==4.9.0.80

# Dev / testing
pytest>=8.0
pytest-cov>=5.0
